linear-progress: move component logic to new file
